[ASV-677][ASV-1211] Removed not necessary classes
