+++ app/src/main/java/cm/aptoide/pt/autoupdate/AutoUpdateViewModel.java
@@ -1,89 +1,66 @@
 package cm.aptoide.pt.autoupdate;
 
 public class AutoUpdateViewModel {
-  private final String packageName;
-  private final int localVersionCode;
-  private String md5;
-  private int vercode;
-  private String infoPackageName;
-  private int appId;
-  private String path;
-  private int mindsdk;
-  private int minAptoideVercode;
 
-  public AutoUpdateViewModel(String packageName, int localVersionCode) {
-    this.packageName = packageName;
-    this.localVersionCode = localVersionCode;
-    this.md5 = "";
-    this.vercode = -1;
-    this.infoPackageName = "";
-    this.appId = -1;
-    this.path = "";
-    this.mindsdk = 0;
-    this.minAptoideVercode = 0;
+  private final long versionCode;
+  private final String uri;
+  private final String md5;
+  private final String minSdk;
+  private final boolean loading;
+  private final Error error;
+
+  public AutoUpdateViewModel(long versionCode, String uri, String md5, String minSdk) {
+    this.versionCode = versionCode;
+    this.uri = uri;
+    this.md5 = md5;
+    this.minSdk = minSdk;
+    loading = false;
+    error = null;
   }
 
-  public String getPackageName() {
-    return packageName;
+  public AutoUpdateViewModel(Error error) {
+    this.error = error;
+    versionCode = -1;
+    uri = null;
+    md5 = null;
+    minSdk = null;
+    loading = false;
   }
 
-  public int getLocalVersionCode() {
-    return localVersionCode;
+  public AutoUpdateViewModel(boolean loading) {
+    this.loading = loading;
+    versionCode = -1;
+    uri = null;
+    md5 = null;
+    minSdk = null;
+    error = null;
+  }
+
+  public long getVersionCode() {
+    return versionCode;
+  }
+
+  public String getUri() {
+    return uri;
   }
 
   public String getMd5() {
     return md5;
   }
 
-  public void setMd5(String md5) {
-    this.md5 = md5;
+  public String getMinSdk() {
+    return minSdk;
   }
 
-  public int getVercode() {
-    return vercode;
+  public boolean isLoading() {
+    return loading;
   }
 
-  public void setVercode(int vercode) {
-    this.vercode = vercode;
+  public Error getError() {
+    return error;
   }
 
-  public String getInfoPackageName() {
-    return infoPackageName;
-  }
-
-  public void setInfoPackageName(String infoPackageName) {
-    this.infoPackageName = infoPackageName;
-  }
-
-  public int getAppId() {
-    return appId;
-  }
-
-  public void setAppId(int appId) {
-    this.appId = appId;
-  }
-
-  public String getPath() {
-    return path;
-  }
-
-  public void setPath(String path) {
-    this.path = path;
-  }
-
-  public int getMindsdk() {
-    return mindsdk;
-  }
-
-  public void setMindsdk(int mindsdk) {
-    this.mindsdk = mindsdk;
-  }
-
-  public int getMinAptoideVercode() {
-    return minAptoideVercode;
-  }
-
-  public void setMinAptoideVercode(int minAptoideVercode) {
-    this.minAptoideVercode = minAptoideVercode;
+  public enum Error {
+    NETWORK, GENERIC
   }
 }
